refactor: components fodler and removing placiholder
