+++ app/(site)/about/page.tsx
@@ -1,7 +1,7 @@
-import NodeAnimate from '@/components/animation/node'
-import { StarIcon } from '@/components/icons/StarIcon'
-import { ExperienceItem } from '@/components/timeline'
-import { ExternalLink } from '@/components/ui/external-link'
+import NodeAnimate from '@/app/_components/animate/node'
+import { StarIcon } from '@/app/_components/icons/StarIcon'
+import { ExperienceItem } from '@/app/_components/timeline'
+import { ExternalLink } from '@/app/_components/external-link'
 import { generateMeta, keywords, getOgImage } from '@/lib/meta'
 import { allExperiences as experiences } from '@/.contentlayer/generated'
 import { Metadata } from 'next'
