+++ app/_components/external-link.tsx
