+++ app/_components/icons/StarIcon.tsx
