+++ app/_components/note/index.ts
