+++ app/_components/note/item.tsx
