+++ app/_components/waving-hand.tsx
