+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css'
 import '../styles/prose.css'
-import { Providers } from '@/components/providers'
+import { Providers } from '@/app/_components/providers'
 import { cn } from '@/lib/utils/cn'
 import { primary, secondary } from '@/lib/fonts'
 import { Analytics } from '@vercel/analytics/react'
