+++ app/note/[...slug]/page.tsx
@@ -1,8 +1,8 @@
 import { allNotes } from '@/.contentlayer/generated'
-import NoteWrapper from '@/components/animation/animation-wrapper'
-import { BackToTop } from '@/components/back-to-top'
-import { Mdx } from '@/components/mdx'
-import { ShareNote } from '@/components/share-note'
+import NoteWrapper from '@/app/_components/animate/wrapper'
+import { BackToTop } from '@/app/_components/back-to-top'
+import { Mdx } from '@/app/_components/mdx'
+import { ShareNote } from '@/app/_components/note/share-note'
 import { generateMeta, getOgImage } from '@/lib/meta'
 import { formatDate } from '@/lib/utils/date'
 import { getNoteLenghtIcon } from '@/lib/utils/note-length'
