--- components/footer.tsx
@@ -1,3 +1,0 @@
-export default function Footer() {
-	return <footer></footer>
-}
