+++ package.json
@@ -9,7 +9,6 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
-		"@plaiceholder/next": "^3.0.0",
 		"@studio-freight/react-lenis": "^0.0.36",
 		"@vercel/analytics": "^1.1.1",
 		"@vercel/speed-insights": "^1.0.2",
@@ -22,7 +21,6 @@
 		"next-contentlayer": "^0.3.4",
 		"next-sanity": "^6.0.1",
 		"next-themes": "^0.2.1",
-		"plaiceholder": "^3.0.0",
 		"react": "^18",
 		"react-dom": "^18",
 		"rehype-autolink-headings": "^7.1.0",
